requirements.txt
connexion[swagger-ui]
pyserial